+++ src/main/java/dev/dubhe/anvilcraft/item/ResonatorItem.java
@@ -1,11 +1,16 @@
 package dev.dubhe.anvilcraft.item;
 
+import com.mojang.blaze3d.vertex.PoseStack;
+import dev.dubhe.anvilcraft.block.multipart.AbstractMultiPartBlock;
+import dev.dubhe.anvilcraft.client.renderer.item.ItemUseAnimationTransform;
+import dev.dubhe.anvilcraft.init.block.ModBlocks;
 import dev.dubhe.anvilcraft.init.enchantment.ModEnchantmentTags;
 import dev.dubhe.anvilcraft.init.item.ModComponents;
 import dev.dubhe.anvilcraft.init.item.ModItemTags;
-import dev.dubhe.anvilcraft.init.item.ModItems;
+import dev.dubhe.anvilcraft.network.ResonanceMiningEffectPacket;
 import net.minecraft.ChatFormatting;
 import net.minecraft.advancements.CriteriaTriggers;
+import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.Holder;
@@ -14,6 +19,7 @@ import net.minecraft.core.HolderSet;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
@@ -22,9 +28,11 @@ import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.EquipmentSlotGroup;
+import net.minecraft.world.entity.HumanoidArm;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
@@ -37,24 +45,32 @@ import net.minecraft.world.item.SwordItem;
 import net.minecraft.world.item.Tier;
 import net.minecraft.world.item.TieredItem;
 import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.UseAnim;
 import net.minecraft.world.item.component.CustomModelData;
 import net.minecraft.world.item.component.ItemAttributeModifiers;
 import net.minecraft.world.item.component.Tool;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.ItemEnchantments;
+import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.gameevent.GameEvent;
+import net.minecraft.world.phys.BlockHitResult;
+import net.minecraft.world.phys.HitResult;
+import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 import net.neoforged.neoforge.common.ItemAbilities;
 import net.neoforged.neoforge.common.ItemAbility;
+import net.neoforged.neoforge.network.PacketDistributor;
 import org.jetbrains.annotations.Range;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.WeakHashMap;
 import java.util.function.Consumer;
 
 public abstract class ResonatorItem extends TieredItem {
@@ -63,6 +79,12 @@ public abstract class ResonatorItem extends TieredItem {
     public static final int SHOVEL_MODE = 2;
     public static final int HOE_MODE = 3;
     public static final int PICKAXE_MODE = 4;
+    private static final int USE_DURATION = 72000;
+    private static final int STANDARD_RESONANCE_MINING_TICKS = 20;
+    private static final int STANDARD_RESONANCE_MINING_DURABILITY_COST = 128;
+
+    private final Map<LivingEntity, MiningTarget> clientMiningTargets = new WeakHashMap<>();
+    private final Map<LivingEntity, MiningTarget> serverMiningTargets = new WeakHashMap<>();
 
     public ResonatorItem(Tier tier, Properties properties) {
         super(
@@ -71,25 +93,42 @@ public abstract class ResonatorItem extends TieredItem {
         );
     }
 
-    private boolean isTranscendence(ItemStack stack) {
-        return stack.is(ModItems.TRANSCENDENCE_RESONATOR);
+    @Override
+    @SuppressWarnings("removal")
+    public void initializeClient(Consumer<IClientItemExtensions> consumer) {
+        consumer.accept(new IClientItemExtensions() {
+            @Override
+            public boolean applyForgeHandTransform(
+                PoseStack poseStack,
+                LocalPlayer player,
+                HumanoidArm arm,
+                ItemStack stack,
+                float partialTick,
+                float equipProgress,
+                float swingProgress
+            ) {
+                return ItemUseAnimationTransform.applyCrossbowCharge(
+                    poseStack,
+                    player,
+                    arm,
+                    stack,
+                    partialTick,
+                    equipProgress,
+                    resonanceMiningTicks()
+                );
+            }
+        });
     }
 
     @Override
     public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag) {
         super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
-
-        if (isTranscendence(stack)) {
-            tooltipComponents.add(
-                Component.translatable("tooltip.anvilcraft.resonator.mining_desc", Component.keybind("key.anvilcraft.switch_tool_mode"))
-                    .withStyle(ChatFormatting.GRAY)
-            );
-        } else {
-            tooltipComponents.add(
-                Component.translatable("tooltip.anvilcraft.resonator.desc", Component.keybind("key.anvilcraft.switch_tool_mode"))
-                    .withStyle(net.minecraft.ChatFormatting.GRAY)
-            );
-        }
+        tooltipComponents.add(
+            Component.translatable(
+                "tooltip.anvilcraft.resonator.mining_desc",
+                Component.keybind("key.anvilcraft.switch_tool_mode")
+            ).withStyle(ChatFormatting.GRAY)
+        );
     }
 
     public static ItemAttributeModifiers createAttributes(Tier tier, float attackDamage, float attackSpeed) {
@@ -213,6 +252,14 @@ public abstract class ResonatorItem extends TieredItem {
 
     protected abstract double getBaseAttackDamage();
 
+    protected int resonanceMiningTicks() {
+        return STANDARD_RESONANCE_MINING_TICKS;
+    }
+
+    protected int resonanceMiningDurabilityCost() {
+        return STANDARD_RESONANCE_MINING_DURABILITY_COST;
+    }
+
     @Override
     public void inventoryTick(ItemStack stack, Level level, Entity entity, int slotId, boolean isSelected) {
         super.inventoryTick(stack, level, entity, slotId, isSelected);
@@ -239,20 +286,219 @@ public abstract class ResonatorItem extends TieredItem {
         return stack.getDamageValue() >= stack.getMaxDamage() - 1;
     }
 
+    private boolean canStartResonanceMining(ItemStack stack) {
+        if (isTooDamagedToUse(stack)) return false;
+        int durabilityCost = resonanceMiningDurabilityCost();
+        return durabilityCost <= 0 || stack.getMaxDamage() - stack.getDamageValue() > durabilityCost;
+    }
+
+    @Override
+    public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand usedHand) {
+        ItemStack stack = player.getItemInHand(usedHand);
+        if (getMode(stack) == AUTO_MODE && canStartResonanceMining(stack)) {
+            return InteractionResultHolder.fail(stack);
+        }
+        return super.use(level, player, usedHand);
+    }
+
+    @Override
+    public InteractionResult onItemUseFirst(ItemStack stack, UseOnContext context) {
+        if (getMode(stack) != AUTO_MODE || !canStartResonanceMining(stack)) {
+            return super.onItemUseFirst(stack, context);
+        }
+        return startResonanceMining(context);
+    }
+
     @Override
     public InteractionResult useOn(UseOnContext context) {
         ItemStack stack = context.getItemInHand();
         int mode = ResonatorItem.getMode(stack);
         return switch (mode) {
-            case AUTO_MODE -> {
-                yield InteractionResult.PASS;
-            }
+            case AUTO_MODE -> canStartResonanceMining(stack)
+                ? startResonanceMining(context)
+                : InteractionResult.PASS;
             case AXE_MODE -> this.useOnAsAxe(context);
             case SHOVEL_MODE -> this.useOnAsShovel(context);
             case HOE_MODE -> this.useOnAsHoe(context);
             case PICKAXE_MODE -> this.useOnAsPickaxe(context);
             default -> super.useOn(context);
         };
+    }
+
+    private InteractionResult startResonanceMining(UseOnContext context) {
+        if (context.getHand() != InteractionHand.MAIN_HAND) return InteractionResult.PASS;
+
+        Level level = context.getLevel();
+        BlockPos pos = context.getClickedPos();
+        if (!canResonanceMine(level.getBlockState(pos), level, pos)) return InteractionResult.PASS;
+
+        Player player = context.getPlayer();
+        if (player == null) return InteractionResult.PASS;
+
+        BlockHitResult hitResult = new BlockHitResult(
+            context.getClickLocation(),
+            context.getClickedFace(),
+            pos.immutable(),
+            context.isInside()
+        );
+        MiningTarget target = new MiningTarget(hitResult, context.getHand(), getEffectPositions(level, pos));
+        miningTargets(level).put(player, target);
+        player.startUsingItem(context.getHand());
+        sendMiningEffects(level, target.effectPositions(), resonanceMiningTicks() + 2);
+        return InteractionResult.CONSUME;
+    }
+
+    @Override
+    public int getUseDuration(ItemStack stack, LivingEntity entity) {
+        return getMode(stack) == AUTO_MODE ? USE_DURATION : super.getUseDuration(stack, entity);
+    }
+
+    @Override
+    public UseAnim getUseAnimation(ItemStack stack) {
+        return getMode(stack) == AUTO_MODE ? UseAnim.CROSSBOW : UseAnim.NONE;
+    }
+
+    public static float resonanceMiningProgress(Level level, Player player, float partialTick) {
+        if (!(player.getUseItem().getItem() instanceof ResonatorItem resonator)) return -1.0F;
+        if (!resonator.miningTargets(level).containsKey(player)) return -1.0F;
+
+        ItemStack stack = player.getUseItem();
+        int elapsedTicks = stack.getUseDuration(player) - player.getUseItemRemainingTicks();
+        return Math.min(1.0F, (elapsedTicks + partialTick) / resonator.resonanceMiningTicks());
+    }
+
+    @Override
+    public void onUseTick(Level level, LivingEntity livingEntity, ItemStack stack, int remainingUseDuration) {
+        Map<LivingEntity, MiningTarget> targets = miningTargets(level);
+        MiningTarget target = targets.get(livingEntity);
+        if (target == null) {
+            if (!level.isClientSide) livingEntity.stopUsingItem();
+            return;
+        }
+
+        BlockHitResult hit = getTargetedBlock(livingEntity);
+        if (hit == null || !target.hitPos().equals(hit.getBlockPos()) || !canStartResonanceMining(stack)) {
+            stopResonanceMining(level, livingEntity, target);
+            return;
+        }
+
+        BlockState state = level.getBlockState(target.hitPos());
+        if (!canResonanceMine(state, level, target.hitPos())) {
+            stopResonanceMining(level, livingEntity, target);
+            return;
+        }
+
+        int elapsedTicks = getUseDuration(stack, livingEntity) - remainingUseDuration;
+        if (!level.isClientSide && elapsedTicks % 3 == 0) {
+            float pitch = 0.75f + 0.04f * elapsedTicks;
+            level.playSound(null, target.hitPos(), SoundEvents.AMETHYST_BLOCK_RESONATE, SoundSource.BLOCKS, 0.8f, pitch);
+        }
+        if (elapsedTicks < resonanceMiningTicks()) return;
+
+        int damageBeforeMining = stack.getDamageValue();
+        boolean destroyed = livingEntity instanceof ServerPlayer player
+            && player.gameMode.destroyBlock(target.hitPos());
+        targets.remove(livingEntity);
+        if (destroyed) {
+            consumeResonanceMiningDurability(stack, damageBeforeMining);
+            level.playSound(null, target.hitPos(), SoundEvents.AMETHYST_CLUSTER_BREAK, SoundSource.BLOCKS, 1.0f, 0.7f);
+        }
+        sendMiningEffects(level, target.effectPositions(), 0);
+        livingEntity.stopUsingItem();
+    }
+
+    @Override
+    public void releaseUsing(ItemStack stack, Level level, LivingEntity livingEntity, int remainingUseDuration) {
+        MiningTarget target = miningTargets(level).remove(livingEntity);
+        if (target == null) return;
+
+        sendMiningEffects(level, target.effectPositions(), 0);
+        int elapsedTicks = getUseDuration(stack, livingEntity) - remainingUseDuration;
+        if (elapsedTicks >= resonanceMiningTicks() || !(livingEntity instanceof Player player)) return;
+
+        BlockHitResult hit = getTargetedBlock(player);
+        if (hit == null || !target.hitPos().equals(hit.getBlockPos())) return;
+        AnvilHammerItem.interactWithBlock(
+            player,
+            target.hitPos(),
+            level,
+            stack,
+            target.hand(),
+            target.hitResult()
+        );
+    }
+
+    private void consumeResonanceMiningDurability(ItemStack stack, int damageBeforeMining) {
+        int durabilityCost = resonanceMiningDurabilityCost();
+        if (durabilityCost <= 0) return;
+        // 直接写入损伤值，避免耐久附魔改变本次消耗。
+        stack.setDamageValue(damageBeforeMining + durabilityCost);
+    }
+
+    private Map<LivingEntity, MiningTarget> miningTargets(Level level) {
+        return level.isClientSide ? this.clientMiningTargets : this.serverMiningTargets;
+    }
+
+    public static boolean isResonanceMining(Level level, Player player, BlockPos pos) {
+        if (!(player.getUseItem().getItem() instanceof ResonatorItem resonator)) return false;
+        MiningTarget target = resonator.miningTargets(level).get(player);
+        return target != null && target.hitPos().equals(pos);
+    }
+
+    private void stopResonanceMining(Level level, LivingEntity livingEntity, MiningTarget target) {
+        miningTargets(level).remove(livingEntity);
+        sendMiningEffects(level, target.effectPositions(), 0);
+        livingEntity.stopUsingItem();
+    }
+
+    private static List<BlockPos> getEffectPositions(Level level, BlockPos hitPos) {
+        BlockState state = level.getBlockState(hitPos);
+        if (state.is(ModBlocks.LARGE_CAKE)) return List.of(hitPos.immutable());
+        if (!(state.getBlock() instanceof AbstractMultiPartBlock<?> multiPartBlock)) {
+            return List.of(hitPos.immutable());
+        }
+
+        return getMultiPartEffectPositions(level, hitPos, state, multiPartBlock);
+    }
+
+    private static <P extends Enum<P>> List<BlockPos> getMultiPartEffectPositions(
+        Level level,
+        BlockPos hitPos,
+        BlockState hitState,
+        AbstractMultiPartBlock<P> multiPartBlock
+    ) {
+        List<BlockPos> positions = new ArrayList<>();
+        for (P part : multiPartBlock.getParts()) {
+            BlockPos partPos = hitPos.offset(multiPartBlock.offsetFrom(hitState, part));
+            if (level.getBlockState(partPos).is(multiPartBlock)) positions.add(partPos.immutable());
+        }
+        return positions.isEmpty() ? List.of(hitPos.immutable()) : List.copyOf(positions);
+    }
+
+    private static void sendMiningEffects(Level level, List<BlockPos> positions, int durationTicks) {
+        for (BlockPos pos : positions) {
+            sendMiningEffect(level, pos, durationTicks);
+        }
+    }
+
+    private static void sendMiningEffect(Level level, BlockPos pos, int durationTicks) {
+        if (!(level instanceof ServerLevel serverLevel)) return;
+        PacketDistributor.sendToPlayersTrackingChunk(
+            serverLevel,
+            new ChunkPos(pos),
+            new ResonanceMiningEffectPacket(pos, durationTicks)
+        );
+    }
+
+    private static BlockHitResult getTargetedBlock(LivingEntity livingEntity) {
+        if (!(livingEntity instanceof Player player)) return null;
+        HitResult hit = player.pick(player.blockInteractionRange(), 0.0f, false);
+        return hit.getType() == HitResult.Type.BLOCK ? (BlockHitResult) hit : null;
+    }
+
+    static boolean canResonanceMine(BlockState state, Level level, BlockPos pos) {
+        if (state.isAir()) return false;
+        return state.getDestroySpeed(level, pos) >= 0.0f;
     }
 
     public InteractionResult useOnAsAxe(UseOnContext context) {
@@ -386,6 +632,16 @@ public abstract class ResonatorItem extends TieredItem {
             case PICKAXE_MODE -> ItemAbilities.DEFAULT_PICKAXE_ACTIONS.contains(itemAbility);
             default -> false;
         };
+    }
+
+    private record MiningTarget(
+        BlockHitResult hitResult,
+        InteractionHand hand,
+        List<BlockPos> effectPositions
+    ) {
+        private BlockPos hitPos() {
+            return hitResult.getBlockPos();
+        }
     }
 
     public static class ResonatorHolder extends Holder.Reference<Item> {
